requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.1
multidict==6.0.3
python-dotenv==0.21.0
pytz==2022.6
SQLAlchemy==1.4.45
Werkzeug==2.2.2
WTForms==3.0.1
yarl==1.8.2
